Move Signature verification and k generation into a single file, and fix aggsig
